(refactor): move Item class to domain package and refactor for loop to enhanced for loop
