+++ Java-Cucumber/src/test/java/com/gildedrose/StepDefinitions.java
@@ -1,5 +1,6 @@
 package com.gildedrose;
 
+import com.gildedrose.domain.Item;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
